# Improve large annotation load and display speed