--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Speed up annotation centroid queries ([#1981](../../pull/1981))
+- Improve large annotation load and display speed ([#1982](../../pull/1982))
 
 ## 1.33.0