--- a/girder_annotation/girder_large_image_annotation/models/annotation.py
+++ b/girder_annotation/girder_large_image_annotation/models/annotation.py
@@ -663,7 +663,7 @@ class Annotation(AccessControlledModel):
         })
 
         self.exposeFields(AccessType.READ, (
-            'annotation', '_version', '_elementQuery', '_active',
+            'annotation', '_version', '_elementQuery', '_active', '_elementCount', '_detailsCount',
         ) + self.baseFields)
         events.bind('model.item.remove', 'large_image_annotation', self._onItemRemove)
         events.bind('model.item.copy.prepare', 'large_image_annotation', self._prepareCopyItem)
@@ -946,7 +946,7 @@ class Annotation(AccessControlledModel):
             annotation['annotation']['name'] = now.strftime('Annotation %Y-%m-%d %H:%M')
 
         def replaceElements(query, doc, *args, **kwargs):
-            Annotationelement().updateElements(doc)
+            elementCount, detailsCount = Annotationelement().updateElements(doc)
             elements = doc['annotation'].pop('elements', None)
             if self._historyEnabled:
                 oldAnnotation = self.collection.find_one(query)
@@ -954,6 +954,9 @@ class Annotation(AccessControlledModel):
                     oldAnnotation['_annotationId'] = oldAnnotation.pop('_id')
                     oldAnnotation['_active'] = False
                     insert_one(oldAnnotation)
+            if elements is not None:
+                doc['_elementCount'] = elementCount
+                doc['_detailsCount'] = detailsCount
             ret = replace_one(query, doc, *args, **kwargs)
             if elements:
                 doc['annotation']['elements'] = elements
@@ -966,10 +969,13 @@ class Annotation(AccessControlledModel):
             # the annotation without elements, then restore the elements.
             doc.setdefault('_id', ObjectId())
             if doc['annotation'].get('elements') is not None:
-                Annotationelement().updateElements(doc)
+                elementCount, detailsCount = Annotationelement().updateElements(doc)
             # If we are inserting, we shouldn't have any old elements, so don't
             # bother removing them.
             elements = doc['annotation'].pop('elements', None)
+            if elements is not None:
+                doc['_elementCount'] = elementCount
+                doc['_detailsCount'] = detailsCount
             ret = insert_one(doc, *args, **kwargs)
             if elements is not None:
                 doc['annotation']['elements'] = elements